Move AuraInputConfig Class to Data Folder
